% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{Score genes based on their proximity to the reference genes.}
\usage{
adjacency(estimate = densest)
}
\arguments{
\item{estimate}{A function that will be used to summarize the distance
values for each gene. See \code{\link[=densest]{densest()}} for the default implementation.}
}
\value{
An object of class \code{geposan_method}.
}
\description{
Score genes based on their proximity to the reference genes.
}
