% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze by applying the specified preset.}
\usage{
analyze(preset)
}
\arguments{
\item{preset}{The preset to use which can be created using \code{\link[=preset]{preset()}}.}
}
\value{
A \link{data.table} with one row for each gene identified by it's ID
(\code{gene} column). The additional columns contain the resulting scores per
method and are named after the method IDs.
}
\description{
Analyze by applying the specified preset.
}
