% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromosomes}
\alias{chromosomes}
\title{Information on chromosomes for each included species.}
\format{
A \link{data.table} with the following columns:
\describe{
\item{species}{Species ID}
\item{id}{Chromosome ID, theses are Ensembl sequence IDs}
\item{name}{Chromosome name}
\item{length}{Length in base pairs}
}
}
\usage{
chromosomes
}
\description{
Information on chromosomes for each included species.
}
\keyword{datasets}
