% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_clustering.R
\name{clusteriness}
\alias{clusteriness}
\title{Perform a cluster analysis.}
\usage{
clusteriness(data, span = 1e+05, weight = 0.7)
}
\arguments{
\item{data}{The values that should be scored.}

\item{span}{The maximum span of values considered to be in one cluster.}

\item{weight}{The weight that will be given to the next largest cluster in
relation to the previous one. For example, if \code{weight} is 0.7 (the
default), the first cluster will weigh 1.0, the second 0.7, the third 0.49
etc.}
}
\value{
A score between 0.0 and 1.0 summarizing how much the data clusters.
}
\description{
This function will cluster the data using \code{\link[stats:hclust]{stats::hclust()}} and
\code{\link[stats:cutree]{stats::cutree()}}. Every cluster with at least two members qualifies for
further analysis. Clusters are then ranked based on their size in relation
to the total number of values. The return value is a final score between
0.0 and 1.0. Lower ranking clusters contribute less to this score.
}
