% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{compare}
\alias{compare}
\title{Compare a set of genes with the ranking.}
\usage{
compare(ranking, comparison_gene_ids)
}
\arguments{
\item{ranking}{A ranking created using \code{\link[=ranking]{ranking()}}.}

\item{comparison_gene_ids}{IDs of the genes of interest.}
}
\value{
A comparison object with the following items:
\describe{
\item{\code{mean}}{The mean score of the genes.}
\item{\code{min}}{The lowest score of the genes.}
\item{\code{max}}{The highest score of the genes.}
\item{\code{mean_rank}}{The mean rank of the genes.}
\item{\code{first_rank}}{The first rank of the genes.}
\item{\code{last_rank}}{The last rank of the genes.}
\item{\code{p_value}}{p-value for the null hypothesis that the comparison
genes do \emph{not} rank better than other genes. In other words: A low
p-value means that the comparison genes rank particularly high.}
}
}
\description{
Compare a set of genes with the ranking.
}
