% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Compute the mean correlation coefficient comparing gene distances with a set
of reference genes.}
\usage{
correlation()
}
\value{
An object of class \code{geposan_method}.
}
\description{
Compute the mean correlation coefficient comparing gene distances with a set
of reference genes.
}
