% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_distance.R
\name{distance}
\alias{distance}
\title{Score the distance of genes to the telomeres across species.}
\usage{
distance(
  id = "distance",
  name = "Distance",
  description = "Distance to telomeres",
  summarize = densest
)
}
\arguments{
\item{id}{Unique ID for the method and its results.}

\item{name}{Human readable name for the method.}

\item{description}{Method description.}

\item{summarize}{A function for combining the different proximities into one
metric. By default, \code{\link[stats:median]{stats::median()}} is used. Other suggested options
include \code{\link[=min]{min()}} and \code{\link[=mean]{mean()}}.}
}
\value{
An object of class \code{geposan_method}.
}
\description{
A score will be given to each gene such that 0.0 corresponds to the maximal
distance across all genes and 1.0 corresponds to a distance of 0.
}
