% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{Information on human genes within the Ensembl database.}
\format{
A \link{data.table} with 60568 rows and 3 variables:
\describe{
\item{id}{Ensembl gene ID}
\item{name}{The gene's HGNC name}
\item{chrosome}{The human chromosome the gene is located on}
\item{n_species}{Number of known species with the gene.}
}
}
\usage{
genes
}
\description{
This includes only genes on the primary suggested assembly of the human
nuclear DNA.
}
\keyword{datasets}
