% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method}
\alias{method}
\title{Describe a new method for analyzing gene position data.}
\usage{
method(id, name, description, help, func)
}
\arguments{
\item{id}{Unique identifier for the method.}

\item{name}{Human readable name.}

\item{description}{Slightly longer description.}

\item{help}{Context help for user interfaces.}

\item{func}{Function to apply the method. The function should accept two
parameters: an object of class \code{geposan_preset} as input and a function to
report progress information to as a numeric value. The return value should
be an object of class \code{geposan_result}.}
}
\value{
An object of class \code{geposan_method}.
}
\description{
Describe a new method for analyzing gene position data.
}
