% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{optimal_weights}
\alias{optimal_weights}
\title{Find the best weights to rank the results.}
\usage{
optimal_weights(analysis, methods, reference_gene_ids, target = "mean")
}
\arguments{
\item{analysis}{Results from \code{\link[=analyze]{analyze()}} or \code{\link[=ranking]{ranking()}}.}

\item{methods}{Methods to include in the score.}

\item{reference_gene_ids}{IDs of the reference genes.}

\item{target}{The optimization target. It may be one of "mean", "median",
"min" or "max" and results in the respective rank being optimized.}
}
\value{
Named list pairing method names with their optimal weights. This
can be used as an argument to \code{\link[=ranking]{ranking()}}.
}
\description{
This function finds the optimal parameters to \code{\link[=ranking]{ranking()}} that result in the
reference genes ranking particulary high.
}
