% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{plot.geposan_ranking}
\alias{plot.geposan_ranking}
\title{S3 method for plotting a ranking.}
\usage{
\method{plot}{geposan_ranking}(ranking, gene_sets = NULL, labels = NULL)
}
\arguments{
\item{gene_sets}{A list of gene sets (containing vectors of gene IDs) that
will be highlighted in the plot.}

\item{labels}{Labels for the gene sets.}
}
\description{
S3 method for plotting a ranking.
}
\seealso{
ranking()
}
