% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_chromosomes}
\alias{plot_chromosomes}
\title{Show the distribution of scores across chromosomes.}
\usage{
plot_chromosomes(ranking)
}
\arguments{
\item{ranking}{The ranking to visualize.}
}
\description{
This function requires the package \code{plotly}.
}
\seealso{
ranking()
}
