% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rankings}
\alias{plot_rankings}
\title{Plot a side-by-side comparison of multiple rankings.}
\usage{
plot_rankings(rankings, gene_sets)
}
\arguments{
\item{rankings}{A named list of rankings to display. The names will be shown
as labels in the plot.}

\item{gene_sets}{A named list of vectors of gene IDs to highlight. The names
will be used to distinguish the sets and in the legend.}
}
\description{
Each ranking's scores will be shown as a vertical violin plot without any
additional markings. The gene sets will be shown as markers on top of the
density visualization.
}
\details{
This function requires the package \code{plotly}.
}
