% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{print.geposan_analysis}
\alias{print.geposan_analysis}
\title{Print an analysis object.}
\usage{
\method{print}{geposan_analysis}(x, ...)
}
\arguments{
\item{x}{The analysis to print.}

\item{...}{Other parameters.}
}
\description{
Print an analysis object.
}
\seealso{
\code{\link[=analyze]{analyze()}}
}
