% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{print.geposan_comparison}
\alias{print.geposan_comparison}
\title{S3 method to print a comparison object.}
\usage{
\method{print}{geposan_comparison}(x, ...)
}
\arguments{
\item{x}{The comparison to print.}

\item{...}{Other parameters.}
}
\description{
S3 method to print a comparison object.
}
\seealso{
\code{\link[=compare]{compare()}}
}
