% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.geposan_method}
\alias{print.geposan_method}
\title{Print a method object.}
\usage{
\method{print}{geposan_method}(x, ...)
}
\arguments{
\item{x}{The method to print.}

\item{...}{Other parameters.}
}
\description{
Print a method object.
}
\seealso{
\code{\link[=method]{method()}}
}
