% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset.R
\name{print.geposan_preset}
\alias{print.geposan_preset}
\title{S3 method to print a preset object.}
\usage{
\method{print}{geposan_preset}(x, ...)
}
\arguments{
\item{x}{The preset to print.}

\item{...}{Other parameters.}
}
\description{
S3 method to print a preset object.
}
\seealso{
\code{\link[=preset]{preset()}}
}
