% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{print.geposan_result}
\alias{print.geposan_result}
\title{Print a result object.}
\usage{
\method{print}{geposan_result}(x, ...)
}
\arguments{
\item{x}{The result to print.}

\item{...}{Other parameters.}
}
\description{
Print a result object.
}
\seealso{
\code{\link[=result]{result()}}
}
