% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{print.geposan_validation}
\alias{print.geposan_validation}
\title{S3 method to print a validation object.}
\usage{
\method{print}{geposan_validation}(x, ...)
}
\arguments{
\item{x}{The validation to print.}

\item{...}{Other parameters.}
}
\description{
S3 method to print a validation object.
}
\seealso{
\code{\link[=validate]{validate()}}
}
