% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking}
\alias{ranking}
\title{Rank the results by computing a score.}
\usage{
ranking(results, weights)
}
\arguments{
\item{results}{Results from \code{\link[=analyze]{analyze()}}.}

\item{weights}{Named list pairing method names with weighting factors.}
}
\description{
This function takes the result from \code{\link[=analyze]{analyze()}} and creates a score by
computing a weighted mean across the different methods' results.
}
