% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result}
\alias{result}
\title{Result of applying a method on gene position data.}
\usage{
result(method_id, scores, details = list())
}
\arguments{
\item{method_id}{ID of the method that produced this result.}

\item{scores}{A \code{data.frame} mapping gene IDs (\code{gene}) to computed scores
between 0.0 and 1.0 (\code{score}).}

\item{details}{Optional details that may contain intermediate results as
well as other information on the method application.}
}
\value{
An object of class \code{geposan_result}.
}
\description{
Result of applying a method on gene position data.
}
