% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Perform cross-validation for the analysis.}
\usage{
validate(analysis, progress = NULL)
}
\arguments{
\item{analysis}{The analysis to validate.}

\item{progress}{An optional progress function that should accept a single
value between 0.0 and 1.0 for progress information.}
}
\value{
An object containing the mean absolute error and the mean percent
rank for the original analysis as well as the validation.
}
\description{
This function reoptimizes the analysis leaving out one of the original
reference genes at a time.
}
