% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{Score genes based on their proximity to the reference genes.}
\usage{
adjacency()
}
\value{
An object of class \code{geposan_method}.
}
\description{
This method finds the distance value with the maximum density for each gene
(i.e. the mode of its estimated distribution). Genes are scored by comparing
those distance values with the values of the reference genes.
}
