% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_adjacency.R
\name{adjacency}
\alias{adjacency}
\title{Score genes based on their proximity to the reference genes.}
\usage{
adjacency(distance_estimate = densest, summarize = min)
}
\arguments{
\item{distance_estimate}{A function that will be used to summarize the
distance values for each gene. See \code{\link[=densest]{densest()}} for the default
implementation.}

\item{summarize}{A function that will be used to combine the different
distances to the reference genes. By default \code{\link[=min]{min()}} is used. That means
the distance to the nearest reference gene will be scored.}
}
\value{
An object of class \code{geposan_method}.
}
\description{
In this case, the distance data that is available for one gene is first
combined. The resulting value is compared to the reference genes and
determines the gene's score in relation to other genes.
}
\seealso{
\code{\link[=species_adjacency]{species_adjacency()}}
}
