% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze by applying the specified preset.}
\usage{
analyze(preset, progress = NULL)
}
\arguments{
\item{preset}{The preset to use which should be created using \code{\link[=preset]{preset()}}.}

\item{progress}{A function to be called for progress information. The
function should accept a number between 0.0 and 1.0 for the current
progress.}
}
\value{
An object containing the results of the analysis with the following
items:
\describe{
\item{\code{preset}}{The preset that was used.}
\item{\code{weights}}{The optimal weights for ranking the reference genes.}
\item{\code{ranking}}{The optimal ranking created using the weights.}
}
}
\description{
Analyze by applying the specified preset.
}
