% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze by applying the specified preset.}
\usage{
analyze(preset, progress = NULL)
}
\arguments{
\item{preset}{The preset to use which should be created using \code{\link[=preset]{preset()}}.}

\item{progress}{A function to be called for progress information. The
function should accept a number between 0.0 and 1.0 for the current
progress.}
}
\value{
An object containing the results of the analysis. It contains a
\link{data.table} with one row for each gene identified by it's ID (\code{gene}
column). The additional columns contain the resulting scores per method
and are named after the method IDs.
}
\description{
Analyze by applying the specified preset.
}
