% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze genes based on position data.}
\usage{
analyze(preset, progress = NULL)
}
\arguments{
\item{preset}{The preset to use which should be created using \code{\link[=preset]{preset()}}.}

\item{progress}{A function to be called for progress information. The
function should accept a number between 0.0 and 1.0 for the current
progress. If no function is provided, a simple text progress bar will be
shown.}
}
\value{
An object containing the results of the analysis with the following
items:
\describe{
\item{\code{preset}}{The preset that was used.}
\item{\code{scores}}{Table containing all scores for each gene.}
\item{\code{results}}{Results from the different methods including details.}
}
}
\description{
Analyze genes based on position data.
}
