% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_clustering.R
\name{clustering}
\alias{clustering}
\title{Process genes clustering their distance to telomeres.}
\usage{
clustering(
  id = "clustering",
  name = "Clustering",
  description = "Clustering of genes",
  span = 5e+05,
  weight = 0.5
)
}
\arguments{
\item{id}{Unique ID for the method and its results.}

\item{name}{Human readable name for the method.}

\item{description}{Method description.}
}
\value{
An object of class \code{geposan_method}.
}
\description{
The result will be cached and can be reused for different presets, because
it is independent of the reference genes in use. Most parameters are exposed
for the \code{\link[=clusteriness]{clusteriness()}} function. See its documentation for more
information.
}
\seealso{
\code{\link[=clusteriness]{clusteriness()}}
}
