% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Process genes clustering their distance to telomeres.}
\usage{
clustering()
}
\value{
An object of class \code{geposan_method}.
}
\description{
The result will be cached and can be reused for different presets, because
it is independent of the reference genes in use.
}
\seealso{
\code{\link[=clusteriness]{clusteriness()}}
}
