% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{compare}
\alias{compare}
\title{Compare a set of genes with the ranking.}
\usage{
compare(ranking, comparison_gene_ids)
}
\arguments{
\item{ranking}{A ranking created using \code{\link[=ranking]{ranking()}}.}

\item{comparison_gene_ids}{IDs of the genes of interest.}
}
\value{
A comparison object with the following items:
\describe{
\item{\code{quantiles}}{A \code{data.table} containing quantile values for the
score, rank and percentile of the comparison genes.
}
\item{\code{mean_score}}{The mean score of the genes.}
\item{\code{mean_rank}}{The mean rank of the genes.}
\item{\code{mean_percentile}}{The mean percentile of the genes.}
\item{\code{test_result}}{Results of applying a Wilcoxon rank sum test.}
}
}
\description{
Compare a set of genes with the ranking.
}
