% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Score genes based on their correlation with the reference genes.}
\usage{
correlation()
}
\value{
An object of class \code{geposan_method}.
}
\description{
Score genes based on their correlation with the reference genes.
}
