% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_correlation.R
\name{correlation}
\alias{correlation}
\title{Score genes based on their correlation with the reference genes.}
\usage{
correlation(summarize = stats::median)
}
\arguments{
\item{summarize}{A function for combining the different correlation
coefficients into one metric. By default, \code{\link[stats:median]{stats::median()}} is used. Other
suggested options include \code{\link[=max]{max()}} and \code{\link[=mean]{mean()}}.}
}
\value{
An object of class \code{geposan_method}.
}
\description{
Score genes based on their correlation with the reference genes.
}
