% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distances}
\alias{distances}
\title{Information on gene positions across species.}
\format{
A \link{data.table} with 1390730 rows and 3 variables:
\describe{
\item{species}{Species ID}
\item{gene}{Gene ID}
\item{distance}{Distance to nearest telomere}
}
}
\usage{
distances
}
\description{
This dataset contains each known value for a gene's distance to the telomeres
per species. The data is sourced from Ensembl.
}
\keyword{datasets}
