% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distances}
\alias{distances}
\title{Information on gene positions across species.}
\format{
A \link{data.table} with the following columns:
\describe{
\item{species}{Species ID}
\item{gene}{Gene ID}
\item{chromosome}{Chromosome ID}
\item{start_position}{Start position in base pairs}
\item{end_position}{End position in base pairs}
\item{distance}{Computed distance to nearest telomere}
}
}
\usage{
distances
}
\description{
This dataset contains each known value for a gene's distance to the telomeres
per species. The data is sourced from Ensembl.
}
\keyword{datasets}
