% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_neural.R
\name{neural}
\alias{neural}
\title{Find genes by training and applying a neural network.}
\usage{
neural(
  id = "neural",
  name = "Neural",
  description = "Assessment by neural network",
  seed = 180199,
  n_models = 5,
  control_ratio = 0.5
)
}
\arguments{
\item{id}{Unique ID for the method and its results.}

\item{name}{Human readable name for the method.}

\item{description}{Method description.}

\item{seed}{The seed will be used to make the results reproducible.}

\item{n_models}{This number specifies how many sets of training data should
be created. For each set, there will be a model trained on the remaining
training data and validated using this set. For non-training genes, the
final score will be the mean of the result of applying the different
models. There should be at least two training sets. The analysis will only
work, if there is at least one reference gene per training set.}

\item{control_ratio}{The proportion of random control genes that is included
in the training data sets in addition to the reference genes. This should
be a numeric value between 0.0 and 1.0.}
}
\value{
An object of class \code{geposan_method}.
}
\description{
Find genes by training and applying a neural network.
}
