% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_positions}
\alias{plot_positions}
\title{Plot gene positions.}
\usage{
plot_positions(species_ids, gene_sets, labels, use_positions = FALSE)
}
\arguments{
\item{species_ids}{IDs of species to show in the plot.}

\item{gene_sets}{A list of gene sets (containing vectors of gene IDs) that
will be highlighted in the plot.}

\item{labels}{Labels for the gene sets. This is required if gene sets are
given and has to have the same length.}

\item{use_positions}{Whether to display positions instead of distances.}
}
\description{
This function requires the package \code{plotly}.
}
