% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scores}
\alias{plot_scores}
\title{Plot a ranking as a scatter plot of scores.}
\usage{
plot_scores(ranking, gene_sets = NULL)
}
\arguments{
\item{ranking}{The ranking to visualize.}

\item{gene_sets}{A named list of gene sets (containing vectors of gene IDs)
that will be highlighted in the plot. The names will be used in the legend.}
}
\description{
This function requires the package \code{plotly}.
}
\seealso{
ranking()
}
