% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scores_by_position}
\alias{plot_scores_by_position}
\title{Plot scores in relation to chromosomal position of genes.}
\usage{
plot_scores_by_position(ranking, chromosome_name, gene_sets = NULL)
}
\arguments{
\item{ranking}{The ranking to visualize.}

\item{chromosome_name}{The chromosome to visualize.}

\item{gene_sets}{Named list of vectors of genes to highlight. The list names
will be used as labels.}
}
\value{
A \code{plotly} figure.
}
\description{
Plot scores in relation to chromosomal position of genes.
}
\seealso{
ranking()
}
