% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset.R
\name{preset}
\alias{preset}
\title{Create a new preset.}
\usage{
preset(
  methods = c("clusteriness", "clusteriness_positions", "correlation",
    "correlation_positions", "neural", "proximity"),
  species_ids = NULL,
  gene_ids = NULL,
  reference_gene_ids = NULL,
  optimization_target = "mean_rank"
)
}
\arguments{
\item{methods}{Methods to apply.}

\item{species_ids}{IDs of species to include.}

\item{gene_ids}{IDs of genes to screen.}

\item{reference_gene_ids}{IDs of reference genes to compare to.}

\item{optimization_target}{Parameter of the reference genes that the ranking
should be optimized for.}
}
\value{
The preset to use with \code{\link[=analyze]{analyze()}}.
}
\description{
A preset is used to specify which methods and inputs should be used for an
analysis. Note that the genes to process should normally include the
reference genes to be able to assess the results later. The genes will be
filtered based on how many species have data for them. Genes which only have
orthologs for less than 25\% of the input species will be excluded from the
preset and the analyis.
}
\details{
Available methods are:
\itemize{
\item \code{clusteriness} How much the gene distances to the nearest telomere
cluster across species.
\item \code{clusteriness_positions} The same as \code{clusteriness} but using absolute
gene positions instead of distances.
\item \code{correlation} The mean correlation of gene distances to the nearest
telomere across species.
\item \code{correlation_positions} Correlation using position data.
\item \code{neural} Assessment by neural network trained using distances.
\item \code{neural_positions} Assessment by neural network trained using absolute
position data.
\item \code{proximity} Mean proximity to telomeres.
}

Available optimization targets are:
\itemize{
\item \code{mean} Mean rank of the reference genes.
\item \code{median} Median rank of the reference genes.
\item \code{max} First rank of the reference genes.
\item \code{min} Last rank of the reference genes.
}
}
