% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset.R
\name{preset}
\alias{preset}
\title{Create a new preset.}
\usage{
preset(
  reference_gene_ids,
  methods = all_methods(),
  species_ids = geposan::species$id,
  gene_ids = geposan::genes$id
)
}
\arguments{
\item{reference_gene_ids}{IDs of reference genes to compare to.}

\item{methods}{List of methods to apply.}

\item{species_ids}{IDs of species to include.}

\item{gene_ids}{IDs of genes to screen.}
}
\value{
The preset to use with \code{\link[=analyze]{analyze()}}.
}
\description{
A preset is used to specify which methods and inputs should be used for an
analysis. Note that the genes to process should normally include the
reference genes to be able to assess the results later. The genes will be
filtered based on how many species have data for them. Genes which only have
orthologs for less than 25\% of the input species will be excluded from the
preset and the analyis. See the different method functions for the available
methods: \code{\link[=clustering]{clustering()}}, \code{\link[=correlation]{correlation()}}, \code{\link[=neural]{neural()}}, \code{\link[=adjacency]{adjacency()}} and
\code{\link[=species_adjacency]{species_adjacency()}}.
}
