% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset.R
\name{preset}
\alias{preset}
\title{Create a new preset.}
\usage{
preset(
  methods = c("clusteriness", "correlation", "neural", "proximity"),
  species_ids = NULL,
  gene_ids = NULL,
  min_n_species = 10,
  reference_gene_ids = NULL
)
}
\arguments{
\item{methods}{Methods to apply.}

\item{species_ids}{IDs of species to include.}

\item{gene_ids}{IDs of genes to screen.}

\item{min_n_species}{Minimum number of orthologs that a gene should have to
be included in the analysis.}

\item{reference_gene_ids}{IDs of reference genes to compare to.}
}
\value{
The preset to use with \code{\link[=analyze]{analyze()}}.
}
\description{
A preset is used to specify which methods and inputs should be used for an
analysis. Note that the genes to process should normally include the
reference genes to be able to assess the results later.
}
\details{
Available methods are:
\itemize{
\item \code{clusteriness} How much the gene distances cluster across species.
\item \code{correlation} The mean correlation with the reference genes.
\item \code{proximity} Mean proximity to telomeres.
\item \code{neural} Assessment by neural network.
}
}
