% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset.R
\name{preset}
\alias{preset}
\title{Create a new preset.}
\usage{
preset(
  reference_gene_ids,
  methods = all_methods(),
  species_ids = geposan::species$id,
  gene_ids = geposan::genes$id,
  species_requirement = 0.25,
  gene_requirement = 0.5
)
}
\arguments{
\item{reference_gene_ids}{IDs of reference genes to compare to.}

\item{methods}{List of methods to apply.}

\item{species_ids}{IDs of species to include.}

\item{gene_ids}{IDs of genes to screen.}

\item{species_requirement}{The proportion of species a gene has to have
orthologs in in order for the gene to qualify.}

\item{gene_requirement}{The proportion of genes that a species has to have
in order for the species to be included in the analysis.}
}
\value{
The preset to use with \code{\link[=analyze]{analyze()}}.
}
\description{
A preset is used to specify which methods and inputs should be used for an
analysis. Note that the genes to process should normally include the
reference genes to be able to assess the results later. The genes will be
filtered based on how many species have data for them. Afterwards, species
that still have many missing genes will also be excluded. See the different
method functions for the available methods: \code{\link[=distance]{distance()}}, \code{\link[=variation]{variation()}},
\code{\link[=clustering]{clustering()}}, \code{\link[=adjacency]{adjacency()}}, \code{\link[=correlation]{correlation()}} and \code{\link[=random_forest]{random_forest()}}.
}
