% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking}
\alias{ranking}
\title{Rank the results by computing a score.}
\usage{
ranking(analysis, weights, min_n_species = 10)
}
\arguments{
\item{analysis}{Analysis object resulting from \code{\link[=analyze]{analyze()}}.}

\item{weights}{Named list pairing method names with weighting factors. Only
methods that are contained within this list will be included.}

\item{min_n_species}{Minimum number of required species per gene. Genes that
have fewer species will not be included in the ranking.}
}
\value{
A ranking object. The object extends the analysis result with
additional columns containing the \code{score} and the \code{rank} of each gene. It
will be ordered by rank.
}
\description{
This function takes the result of \code{\link[=analyze]{analyze()}} and creates a score by
computing a weighted mean across the different methods' results.
}
