% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Information on included species from the Ensembl database.}
\format{
A \link{data.table} with the following columns:
\describe{
\item{id}{Unique species ID}
\item{name}{Human readable species name}
\item{n_chromosomes}{Number of chromosomes}
\item{median_chromosome_length}{Median length of chromosomes}
}
}
\usage{
species
}
\description{
Information on included species from the Ensembl database.
}
\keyword{datasets}
