% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_species_adjacency.R
\name{species_adjacency}
\alias{species_adjacency}
\title{Score genes based on their adjacency to the reference genes within species.}
\usage{
species_adjacency(distance_estimate = min, summarize = stats::median)
}
\arguments{
\item{distance_estimate}{Function for combining the distance differences
within one species.}

\item{summarize}{Function for summarizing the distance values across species.}
}
\value{
An object of class \code{geposan_method}.
}
\description{
For each gene and species, the method will first combine the gene's distances
to the reference genes within that species. Afterwards, the results are
summarized across species and determine the gene's score.
}
\seealso{
\code{\link[=adjacency]{adjacency()}}
}
